Authorizers combined

+++ onyx-backend/budget/src/Budget.Application/Subcategories/RemoveSubcategory/RemoveSubcategoryCommandHandler.cs
@@ -1,10 +1,10 @@
 ﻿using Abstractions.Messaging;
 using Budget.Application.Subcategories.Validator;
+using Budget.Domain.Budgets;
 using Budget.Domain.Categories;
 using Budget.Domain.Subcategories;
 using Budget.Domain.Transactions;
 using Models.Responses;
-using System.ComponentModel.DataAnnotations;
 
 namespace Budget.Application.Subcategories.RemoveSubcategory;
 
@@ -13,18 +13,21 @@ internal sealed class RemoveSubcategoryCommandHandler : ICommandHandler<RemoveSu
     private readonly ISubcategoryRepository _subcategoryRepository;
     private readonly ICategoryRepository _categoryRepository;
     private readonly ITransactionRepository _transactionRepository;
+    private readonly IBudgetRepository _budgetRepository;
     private readonly SubcategoryGlobalValidator _validator;
 
     public RemoveSubcategoryCommandHandler(
         ISubcategoryRepository subcategoryRepository,
         ICategoryRepository categoryRepository,
         ITransactionRepository transactionRepository,
-        SubcategoryGlobalValidator validator)
+        SubcategoryGlobalValidator validator,
+        IBudgetRepository budgetRepository)
     {
         _subcategoryRepository = subcategoryRepository;
         _categoryRepository = categoryRepository;
         _transactionRepository = transactionRepository;
         _validator = validator;
+        _budgetRepository = budgetRepository;
     }
 
     public async Task<Result> Handle(RemoveSubcategoryCommand request, CancellationToken cancellationToken)
@@ -68,10 +71,36 @@ internal sealed class RemoveSubcategoryCommandHandler : ICommandHandler<RemoveSu
         var relatedTransactions = relatedTransactionsGetResult.Value.ToList();
         var category = categoryGetResult.Value;
 
+        var budgetGetResult = await _budgetRepository.GetCurrentBudgetAsync(cancellationToken);
+
+        if (budgetGetResult.IsFailure)
+        {
+            return budgetGetResult.Error;
+        }
+
+        var budget = budgetGetResult.Value;
+
+        if (budget.UnknownSubcategoryId is null)
+        {
+            return RemoveSubcategoryErrors.UnknownSubcategoryNotFoundError;
+        }
+
+        var unknownSubcategoryGetResult = await _subcategoryRepository.GetByIdAsync(
+            budget.UnknownSubcategoryId,
+            cancellationToken);
+
+        if (unknownSubcategoryGetResult.IsFailure)
+        {
+            return unknownSubcategoryGetResult.Error;
+        }
+
+        var unknownSubcategory = unknownSubcategoryGetResult.Value;
+
         var subcategoryServiceRemoveResult = SubcategoryService.RemoveSubcategory(
             subcategory, 
             category, 
-            relatedTransactions);
+            relatedTransactions,
+            unknownSubcategory);
 
         if (subcategoryServiceRemoveResult.IsFailure)
         {
@@ -98,6 +127,14 @@ internal sealed class RemoveSubcategoryCommandHandler : ICommandHandler<RemoveSu
         if (subcategoryRemoveResult.IsFailure)
         {
             return subcategoryRemoveResult.Error;
+        }
+
+        var unknownSubcategoryUpdateResult =
+            await _subcategoryRepository.UpdateAsync(unknownSubcategory, cancellationToken);
+
+        if (unknownSubcategoryUpdateResult.IsFailure)
+        {
+            return unknownSubcategoryUpdateResult.Error;
         }
 
         return Result.Success();
